% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp.R
\name{ppp}
\alias{ppp}
\title{Project prioritization protocol}
\usage{
ppp(x, tree, budget, cost_column_name = "cost",
  success_column_name = "success", locked_in_column_name = NULL,
  locked_out_column_name = NULL, gap = 1e-06, time_limit = NULL,
  threads = 1L, verbose = TRUE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted. Defaults to \code{"cost"}.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted. Defaults to \code{"success"}.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which projects should be locked into the funding scheme. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which projects should be locked out of the funding scheme. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{gap}{\code{numeric} optimality gap. Defaults to 0.000001, so that
solutions will be within 0.0001 % of optimality. No missing values are
permitted.}

\item{time_limit}{\code{numeric} maximum number of seconds that should be
spent searching for a solution after formatting the data. Defaults
to \code{NULL} such that no time limit is imposed.}
}
\value{
A \code{\link[tibble]{tibble}} object that contains the data in
  the argument to \code{x}, with an additional column named
  \code{"solution"} that contains \code{logical} (i.e.
  \code{TRUE}/\code{FALSE} values) which indicate if each project is
  selected for funding in the solution. This object also has the
  following attributes: \code{"objective"} containing the solution's
  objective, \code{"runtime"} denoting the number of seconds that elapsed
  while solving the problem, and \code{"status"} describing the status
  of the solver (e.g. \code{"TODO"} indicates that the
  optimal solution was found, and \code{"TODO"} indicates that the solution
  meets the optimality gap.
}
\description{
Prioritize funding for conservation projects. To identify optimal funding
schemes, the IBM ILOG CPLEX optimization software suite and the
\pkg{cplexAPI} package need to be installed.
}
\details{
TODO
}
