% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_exact_phylo_solution.R
\name{ppp_exact_phylo_solution}
\alias{ppp_exact_phylo_solution}
\title{Prioritize conservation projects with phylogenetic data using exact
algorithms}
\usage{
ppp_exact_phylo_solution(x, y, tree, budget, project_column_name,
  success_column_name, action_column_name, cost_column_name,
  locked_in_column_name = NULL, locked_out_column_name = NULL,
  gap = 1e-06, threads = 1L, number_solutions = 1L,
  time_limit = .Machine$integer.max, number_approx_points = 300,
  verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain data that correspond to
each project. This object should contain data that denote (i)
the name of each project (specified in the argument to
\code{project_column_name}), (ii) the
probability that each project will succeed if all of its actions are funded
(specified in the argument to \code{success_column_name}), (iii)
the enhanced probability that each species will persist if it
is funded (specified as the tip labels in the argument to \code{tree}),
and (iv) and which actions are associated with which projects
(specified in the action names in the argument to \code{y}).
To account for the combined benefits of multiple actions (e.g. baiting
and trapping different invasive species in the same area), additional
projects should be created that indicate the combined cost and
corresponding species' persistence probabilities. Furthermore, this object
must have a baseline project, with a zero cost, that represents the
probability that each species will persist if no other conservation
project is funded.}

\item{y}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing the action data. Here, each row should correspond to a
different action and columns should contain data that correspond to
each action. This object should contain data that denote (i)
the name of each action (specified in the argument to
\code{action_column_name}), (ii) the cost of each action (specified in the
argument to \code{cost_column_name}). If certain actions should be locked
in or out of the solution, then this object should also contain data
that denote (iii) which actions should be locked in (specified using
the argument to \code{locked_in_column_name} if relevant) and (iv) which
actions should be locked out (specified using the argument to
\code{locked_out_column_name} if relevant).}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. This argument corresponds to the
argument to \code{x}. Note that the project names must not contain any
duplicates or missing values.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed. This argument
corresponds to the argument to \code{x}. This column must have
\code{numeric} values which lay between zero and one. No missing values
are permitted.}

\item{action_column_name}{\code{character} name of column that contains
the name for each conservation action. This argument corresponds to the
argument to \code{y}. Note that the project names must not contain any
duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each action. This argument corresponds
to the argument to \code{y}. This column must have \code{numeric} values
which are equal to or greater than zero. No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which actions should be locked into the funding scheme.
This argument corresponds to the argument to \code{y}. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which actions should be locked out of the funding scheme.
This argument corresponds to the argument to \code{y}. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{gap}{\code{numeric} optimality gap. This gap should be expressed as
a proportion. For example, to find a solution that is within 10 % of
optimality, then \code{0.1} should be supplied. No missing values are
permitted. Defaults to \code{0}, so that the optimal solution will be
returned.}

\item{threads}{\code{numeric} number of threads for computational processing.
No missing values are permitted. Defaults to \code{1}.}

\item{number_solutions}{\code{numeric} number of solutions to return. If
the argument is greater than \code{1}, then the output will contain the
set number of solutions that are closest to optimality. No missing values
are permitted. Defaults to \code{1}.}

\item{time_limit}{\code{numeric} maximum number of seconds that should be
spent searching for a solution after formatting the data. Effectively,
defaults to no time limit (but specifically is
\code{.Machine$integer.max}). No missing values are permitted.}

\item{number_approx_points}{\code{numeric} number of points to use for
approximating the probability that branches will go extent. Larger
values increase the precision of these calculations.
No missing values are permitted. Defaults to \code{300}.}

\item{verbose}{\code{logical} should information be printed while solving
the problem? No missing values are permitted. Defaults to \code{FALSE}.}
}
\value{
A \code{\link[tibble]{tibble}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{"solution"}}{\code{integer} solution identifier.}

   \item{\code{"method"}}{\code{character} name of method used to produce the
     solution(s).)}

   \item{\code{"budget"}}{\code{numeric} budget used for generating each of
     the of the solution(s).}

   \item{\code{"obj"}}{\code{numeric} objective value. If phylogenetic data,
     were input, then this column contains the expected phylogenetic
     diversity (Faith 2008) associated with each of the solutions.
     Otherwise, this column contains the sum of the species' weighted
     persistence probabilities .}

   \item{\code{"cost"}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{"optimal"}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal (\code{TRUE}) or not (\code{FALSE}.
     Missing values (\code{NA}) indicate that optimality is unknown
     (i.e. because the method used to produce the solution(s) does not
     provide any bounds on their quality).}

 }
}
\description{
Prioritize funding for conservation projects with phylogenetic data and
using exact algorithms. Unlike other algorithms for solving the
'Project Prioritization Protocol'
(Joseph, Maloney & Possingham 2009), this method can identify
solutions that are guaranteed to be optimal (or within a pre-specified
optimality gap; see Underhill 1994; Rodrigues & Gaston 2002).
\strong{As a consequence, it is strongly recommended to use
this method for developing project prioritizations.}
}
\details{
This function works by formulating the 'Project Prioritization
  Protocol' as a mixed integer programming problem (MIP) and solving it
  using the
  \href{https://www.gurobi.com}{Gurobi optimization software suite}.
  Although \href{https://www.gurobi.com}{Gurobi} is a commercial software,
  academics can obtain a \href{https://user.gurobi.com/download/licenses/free-academic}{special license for no cost}.
   After downloading and installing the
  href{https://www.gurobi.com}{Gurobi} software suite, the
  \pkg{gurobi} package will also need to be installed (see instructions for
  \href{http://www.gurobi.com/documentation/8.1/quickstart_linux/software_installation_guid.html}{Linux},
  \href{http://www.gurobi.com/documentation/8.1/quickstart_mac/software_installation_guid.html}{Mac OSX}, and
  \href{http://www.gurobi.com/documentation/8.1/quickstart_windows/software_installation_guid.html}{Windows} operating systems).
Finally, the \pkg{gurobi} package will also need to be installed (see
instructions for
 \href{http://www.gurobi.com/documentation/8.1/quickstart_linux/r_installing_the_r_package.html}{Linux},
  \href{http://www.gurobi.com/documentation/8.1/quickstart_mac/r_installing_the_r_package.html}{Mac OSX}, and
  \href{http://www.gurobi.com/documentation/8.1/quickstart_windows/r_installing_the_r_package.html}{Windows} operating systems).

The objective of this problem is to maximize the amount of evolutionary
history that is expected to remain within a specified period of time (e.g.
100 years; i.e. 'expected phylogenetic diversity'; Faith 2008).
Let \eqn{I} represent the set of conservation actions (indexed by \eqn{i}).
Let \eqn{C_i} denote the cost for funding action \eqn{i}, and let \eqn{m}
define the maximum expenditure (budget) for funding all the actions. Also,
let \eqn{S} represent each species (indexed by \eqn{s}). To describe the
evolutionary relationships between the species \eqn{s \in S}{s in S},
consider a phylogenetic tree that contains species \eqn{s \in S}{s in S}
with branches of known lengths. This tree can be described using
mathematical notation by letting \eqn{B} represent the branches (indexed by
\eqn{b}) with lengths \eqn{L_b} and letting \eqn{T_{bs}} indicate which
species \eqn{s \in S}{s in S} are associated with which phylogenetic
branches \eqn{b \in B}{b in B} using zeros and ones. Ideally, the set of
species \eqn{S} would contain all of the species in the study
area---including non-threatened species and species which are not associated
with any conservation projects---to fully account for the benefits for
funding different projects.

To guide the prioritization, the conservation actions are organized into
conservation projects. Let \eqn{J} denote the set of conservation projects
(indexed by \eqn{j}, and let \eqn{A_{ij}} denote which actions
\eqn{i \in I}{i in I} comprise each conservation project
\eqn{j \in J}{j in J} using zeros and ones. Next, let \eqn{P_i} represent
the probability of project \eqn{i} being successful if it is funded. Also,
let \eqn{B_{sj}} denote the enhanced probability that each species'
\eqn{s \in S}{s in S} associated with the project \eqn{j \in J}{j in J}
will persist if all of the actions that comprise project \eqn{j} are funded
and that project is allocated to species \eqn{s}.

The binary control variables \eqn{X_i} in this problem indicate whether each
project \eqn{i \in I}{i in I} is funded (variable equal to one) or not
(variable equal to zero). The decision variables in this problem are the
\eqn{Y_{j}}, \eqn{Z_{sj}}, \eqn{E_s}, and \eqn{R_b} variables.
Specifically, the binary \eqn{Y_{j}} variables indicate if project \eqn{j}
is funded (variable equal to one) or not (variable equal to zero) based on
which actions are funded; the binary \eqn{Z_{sj}} variables indicate
if project \eqn{j} is used to manage species \eqn{s} (variable equal to one)
or not (variable equal to zero); the semi-continuous \eqn{E_s} variables
denote the probability that species \eqn{s} will go extinct; and the
semi-continuous \eqn{R_b} variables denote the probability that phylogenetic
branch \eqn{b} will remain.

Now that we have defined all the data and variables, we can formulate
the problem. For convenience, let the symbol used to denote each set also
represent its cardinality (e.g. if there are ten species, let \eqn{S}
represent the set of ten species and also the number ten).

\deqn{
  \mathrm{Maximize} \space \sum_{b = 0}^{B} L_b R_b \space
  \mathrm{(eqn \space 1a)} \\
  \mathrm{Subject \space to} \space R_b = 1 - \prod_{s = 0}^{S}
  ifelse(T_{bs} == 1, \space E_s, \space
  1) \space \forall \space b \in B \space \mathrm{(eqn \space 1b)} \\
  E_s = 1 - \sum_{i = 0}^{I} Y_{is} P_i B_{is} \space \forall \space s \in S
  \space \mathrm{(eqn \space 1c)} \\
  \sum_{i = 0}^{I} C_i \leq m \space
  \mathrm{(eqn \space 1d)} \\
  Z_{sj} \leq Y_{j} \space \forall \space j \in J \space \mathrm{(eqn \space
  1e)} \\
  \sum_{j = 0}^{J} Z_{sj} = 1 \space \forall \space s \in s \space
  \mathrm{eqn \space 1f} \\
  A_{ij} Y_{j} \leq X_{i} \space \forall \space i \in I, j \in J \space
  \mathrm{(eqn \space 1g)} \\
  R_{b} \geq 0, R_{b} \leq 1 \space \forall \space b \in B \space
  \mathrm{(eqn \space 1h)} \\
  E_{s} \geq 0, E_{s} \leq 1 \space \forall \space b \in B \space
  \mathrm{(eqn \space 1i)} \\
  X_{i}, Y_{j}, Z_{sj} \in [0, 1] space \forall \space i \in I, j \in J, s
  \in S \space \mathrm{(eqn \space 1j)}{
  Maximize sum_b^B L_b R_b (eqn 1a); Subject to:
  R_b = 1 - prod_s^S ifelse(T_{bs} == 1, E_s, 1) for all b in B (eqn 1b),
  E_s = 1 - sum_i^I Y_{is} P_i B_{is} for all s in S (eqn 1c),
  sum_i^I C_i <= m (eqn 1d),
  Z_{sj} <= Y_j for all j in J (eqn 1e),
  sum_j^J Z_{sj} = 1 for all s in s (eqn 1f),
  A_{ij} Y_{j} <= X_{i} for all i I, j in J (eqn 1g),
  R_b >= 0, R_b >= 1 for all b in B (eqn 1h),
  E_s >= 0, E_s >= 1 for all s in S (eqn 1i),
  X_i, Y_j, Z_{sj} in [0, 1] for all i in I, j in J, s in S (eqn 1j)}}

The objective (eqn 1a) is to maximize the amount of expected phylogenetic
history that will remain in the future. This is expressed as the sum of
branch lengths (\eqn{L_b}) weighted by the probability that at least one of
the species connected to this branch will not go extinct (\eqn{R_b}).
Constraints (eqn 1b) state that the probability that a branch will remain
(\eqn{R_b}) is equal to one minus the probability that all species will go
extinct. Constraints (eqn 1c) calculate the probability that each species
will go extinct according to their allocated project. Constraint (eqn 1d)
ensures that the cost of all the funded actions do not exceed the budget.
Constraints (eqn 1e) ensure that species can only be allocated to projects
that have all of their actions funded. Constraints (eqn 1f) state that each
species can only be allocated to a single project. Constraints (eqn 1g)
ensure that a project cannot be funded unless all of its actions are funded.
Constraints (eqns 1h, 1i) ensure that the probability variables (\eqn{R_b},
\eqn{E_s}) remain between zero and one. Constraints (eqns 1h) ensure that
the action funding (\eqn{X_j}), project funding (\eqn{Y_j}), and project
allocation (\eqn{Z_{sj}}) variables are binary.

Although this formulation is a mixed integer quadratically constrained
programming problem (due to eqn 1b), it can be linearized and then solved
using commercial mixed integer programming solvers (e.g. Gurobi). This can
be achieved by substituting the product of the species' extinction
probabilities (eqn 1b) with the sum of the log species' extinction
probabilities and using piecewise linear approximations (described in
Hillier & Price 2005 pp. 390--392) to approximate the exponent of this term.
Although this means the problem can only be solved to a pre-specified level
of precision (controlled via the argument to \code{number_approx_points}),
advances in exact algorithm solvers mean that the problem can be solved to a
sufficient degree of precision (e.g. \eqn{1 \times 10^{-5}}{1e-5}) in a
trivial period of time.
}
\examples{
# load built-in data
data(sim_project_data, sim_action_data, sim_tree)

# print simulated project data set
print(sim_project_data)

# print simulated action data
print(sim_action_data)

# print simulated phylogenetic tree data set
print(sim_tree)

# plot the simulated phylogeny
plot(sim_tree, main = "simulated phylogeny")
\donttest{
# verify if guorbi package is installed
if (!require(gurobi, quietly = TRUE))
 stop("the gurobi R package is not installed.")

# find a solution that meets a budget of 300
s1 <- ppp_exact_phylo_solution(sim_project_data, sim_action_data, sim_tree,
                               300, "name", "success", "name", "cost")

# print solution
print(s1)

# print the names of which projects were funded
print(names(s1)[which(unlist(s1[1, sim_action_data$name]))])

# plot solution
ppp_plot_phylo_solution(sim_project_data, sim_action_data, sim_tree, s1,
                        "name", "success", "name", "cost")

# find a solution that meets a budget of 300 and allocates
# funding for the "S1_project" project. For instance, species "S1" might
# be an iconic species that has cultural and economic importance.
sim_action_data2 <- sim_action_data
sim_action_data2$locked_in <- sim_action_data2$name == "S1_project"
s2 <- ppp_exact_phylo_solution(sim_project_data, sim_action_data2, sim_tree,
                               300, "name", "success", "name", "cost",
                               locked_in_column_name = "locked_in")

# print solution
print(s2)

# plot solution
ppp_plot_phylo_solution(sim_project_data, sim_action_data2, sim_tree, s2,
                        "name", "success", "name", "cost")

# find a solution that meets a budget of 300 and does not allocate
# funding for the "S2_project" project. For instance, species "S2"
# might have very little cultural or economic importance. Broadly speaking,
# though, it is better to "lock in" "important" species rather than
# "lock out" unimportant species.
sim_action_data3 <- sim_action_data
sim_action_data3$locked_out <- sim_action_data3$name == "S2_project"
s3 <- ppp_exact_phylo_solution(sim_project_data, sim_action_data3, sim_tree,
                               300, "name", "success", "name", "cost",
                               locked_out_column_name = "locked_out")

# print solution
print(s3)

# plot solution
ppp_plot_phylo_solution(sim_project_data, sim_action_data3, sim_tree, s3,
                        "name", "success", "name", "cost")
}
}
\references{
Faith DP (2008) Threatened species and the potential loss of
phylogenetic diversity: conservation scenarios based on estimated extinction
probabilities and phylogenetic risk analysis. \emph{Conservation Biology},
\strong{22}: 1461--1470.

Hillier FS & Price CC (2005) \emph{International series in operations
research & management science}. Springer.

Joseph LN, Maloney RF & Possingham HP (2009) Optimal allocation of
resources among threatened species: A project prioritization protocol.
\emph{Conservation Biology}, \strong{23}, 328--338.

Rodrigues AS & Gaston KJ (2002) Optimisation in reserve selection
procedures---why not? \emph{Biological Conservation}, \strong{107}: 123-129.

Underhill LG (1994) Optimal and suboptimal reserve selection
algorithms. \emph{Biological Conservation}, \strong{70}: 85--87.
}
\seealso{
For other methods for solving the 'Project Prioritization Protocol'
  problem, see \code{\link{ppp_heuristic_phylo_solution}},
  \code{\link{ppp_manual_phylo_solution}}, and
  \code{\link{ppp_random_phylo_solution}}.
  To visualize the effectiveness of a particular solution, see
  \code{\link{ppp_plot_phylo_solution}}.
}
