% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_exact_spp_solution.R
\name{ppp_exact_spp_solution}
\alias{ppp_exact_spp_solution}
\title{Prioritize conservation projects by maximizing expected weighted species
richness using exact algorithms}
\usage{
ppp_exact_spp_solution(x, y, spp, budget, project_column_name,
  success_column_name, action_column_name, cost_column_name,
  species_column_name, weight_column_name, locked_in_column_name = NULL,
  locked_out_column_name = NULL, gap = 1e-06, threads = 1L,
  number_solutions = 1L, time_limit = .Machine$integer.max,
  number_approx_points = 300, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain data that correspond to
each project. This object should contain data that denote (i)
the name of each project (specified in the argument to
\code{project_column_name}), (ii) the
probability that each project will succeed if all of its actions are funded
(specified in the argument to \code{success_column_name}), (iii)
the enhanced probability that each species will persist if it
is funded,
and (iv) and which actions are associated with which projects
(specified in the action names in the argument to \code{y}).
To account for the combined benefits of multiple actions (e.g. baiting
and trapping different invasive species in the same area), additional
projects should be created that indicate the combined cost and
corresponding species' persistence probabilities. Furthermore, this object
must have a baseline project, with a zero cost, that represents the
probability that each species will persist if no other conservation
project is funded.}

\item{y}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing the action data. Here, each row should correspond to a
different action and columns should contain data that correspond to
each action. This object should contain data that denote (i)
the name of each action (specified in the argument to
\code{action_column_name}), (ii) the cost of each action (specified in the
argument to \code{cost_column_name}). If certain actions should be locked
in or out of the solution, then this object should also contain data
that denote (iii) which actions should be locked in (specified using
the argument to \code{locked_in_column_name} if relevant) and (iv) which
actions should be locked out (specified using the argument to
\code{locked_out_column_name} if relevant).}

\item{spp}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing the species data. Here, each row should correspond
to a different species and columns should contain data that correspond
to each species. This object should contain data that denote
(i) the name of each species (specified in the argument to
\code{species_column_name}). It may also contain (ii) the weight for each
species (specified in the argument to \code{weight_column_name} if
relevant).}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation actions.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. This argument corresponds to the
argument to \code{x}. Note that the project names must not contain any
duplicates or missing values.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed. This argument
corresponds to the argument to \code{x}. This column must have
\code{numeric} values which lay between zero and one. No missing values
are permitted.}

\item{action_column_name}{\code{character} name of column that contains
the name for each conservation action. This argument corresponds to the
argument to \code{y}. Note that the project names must not contain any
duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each action. This argument corresponds
to the argument to \code{y}. This column must have \code{numeric} values
which are equal to or greater than zero. No missing values are permitted.}

\item{species_column_name}{\code{character} name of the column that contains
the name for each species. This argument corresponds to the argument to
\code{spp}.}

\item{weight_column_name}{\code{character} name of the column that contains
the weight for each species. This argument corresponds to the argument to
\code{spp}. This argument defaults to \code{NULL}, such that all species
are assigned an equal weighting.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which actions should be locked into the funding scheme.
This argument corresponds to the argument to \code{y}. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which actions should be locked out of the funding scheme.
This argument corresponds to the argument to \code{y}. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{gap}{\code{numeric} optimality gap. This gap should be expressed as
a proportion. For example, to find a solution that is within 10 \% of
optimality, then \code{0.1} should be supplied. No missing values are
permitted. Defaults to \code{0}, so that the optimal solution will be
returned.}

\item{threads}{\code{numeric} number of threads for computational processing.
No missing values are permitted. Defaults to \code{1}.}

\item{number_solutions}{\code{numeric} number of solutions to return. If
the argument is greater than \code{1}, then the output will contain the
set number of solutions that are closest to optimality. No missing values
are permitted. Defaults to \code{1}.}

\item{time_limit}{\code{numeric} maximum number of seconds that should be
spent searching for a solution after formatting the data. Effectively,
defaults to no time limit (but specifically is
\code{.Machine$integer.max}). No missing values are permitted.}

\item{number_approx_points}{\code{numeric} number of points to use for
approximating the probability that branches will go extent. Larger
values increase the precision of these calculations.
No missing values are permitted. Defaults to \code{300}.}

\item{verbose}{\code{logical} should information be printed while solving
the problem? No missing values are permitted. Defaults to \code{FALSE}.}
}
\value{
A \code{\link[tibble]{tibble}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{"solution"}}{\code{integer} solution identifier.}

   \item{\code{"method"}}{\code{character} name of method used to produce the
     solution(s).)}

   \item{\code{"budget"}}{\code{numeric} budget used for generating each of
     the of the solution(s).}

   \item{\code{"obj"}}{\code{numeric} objective value. If phylogenetic data
     were input, then this column contains the expected phylogenetic
     diversity (Faith 2008) associated with each of the solutions.
     Otherwise, this column contains the expected weighted species richness
     (i.e. the sum of the product between the species' persistence
     probabilities and their weights.}

   \item{\code{"cost"}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{"optimal"}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal (\code{TRUE}) or not (\code{FALSE}).
     Missing values (\code{NA}) indicate that optimality is unknown
     (i.e. because the method used to produce the solution(s) does not
     provide any bounds on their quality).}

 }
}
\description{
Prioritize funding for conservation projects with species weights and
using exact algorithms. Unlike other algorithms for solving the
'Project Prioritization Protocol'
(Joseph, Maloney & Possingham 2009), this method can identify
solutions that are guaranteed to be optimal (or within a pre-specified
optimality gap; see Underhill 1994; Rodrigues & Gaston 2002).
\strong{As a consequence, it is strongly recommended to use
this method for developing project prioritizations.}
}
\details{
This function works by formulating the 'Project Prioritization
  Protocol' as a mixed integer programming problem (MIP) and solving it
  using the
  \href{https://www.gurobi.com}{Gurobi optimization software suite}.
  Although \href{https://www.gurobi.com}{Gurobi} is a commercial software,
  academics can obtain a \href{https://user.gurobi.com/download/licenses/free-academic}{special license for no cost}.
   After downloading and installing the
  href{https://www.gurobi.com}{Gurobi} software suite, the
  \pkg{gurobi} package will also need to be installed (see instructions for
  \href{http://www.gurobi.com/documentation/8.1/quickstart_linux/software_installation_guid.html}{Linux},
  \href{http://www.gurobi.com/documentation/8.1/quickstart_mac/software_installation_guid.html}{Mac OSX}, and
  \href{http://www.gurobi.com/documentation/8.1/quickstart_windows/software_installation_guid.html}{Windows} operating systems).
Finally, the \pkg{gurobi} package will also need to be installed (see
instructions for
 \href{http://www.gurobi.com/documentation/8.1/quickstart_linux/r_installing_the_r_package.html}{Linux},
  \href{http://www.gurobi.com/documentation/8.1/quickstart_mac/r_installing_the_r_package.html}{Mac OSX}, and
  \href{http://www.gurobi.com/documentation/8.1/quickstart_windows/r_installing_the_r_package.html}{Windows} operating systems).

  The mathematical formulation is described in
  \code{\link{ppp_exact_phylo_solution}}. Instead of using a "real"
  phylogeny, it simply uses a star phylogeny---constructed using the species'
  weights---to denote the relationships and importance for conserving
  different species.
}
\examples{
# load built-in data
data(sim_project_data, sim_action_data, sim_species_data)

# print simulated project data set
print(sim_project_data)

# print simulated action data
print(sim_action_data)

# print simulated species data
print(sim_species_data)
\donttest{
# verify if guorbi package is installed
if (!require(gurobi, quietly = TRUE))
 stop("the gurobi R package is not installed.")

# find a solution that meets a budget of 300
s1 <- ppp_exact_spp_solution(sim_project_data, sim_action_data,
                             sim_species_data, 300, "name", "success",
                             "name", "cost", "name", "weight")

# print solution
print(s1)

# print the names of which projects were funded
print(names(s1)[which(unlist(s1[1, sim_action_data$name]))])

# plot solution
ppp_plot_spp_solution(sim_project_data, sim_action_data, sim_species_data,
                      "name", "success", "name", "cost", "name", "weight")

# find a solution that meets a budget of 300 and allocates
# funding for the "S1_project" project. For instance, species "S1" might
# be an iconic species that has cultural and economic importance.
sim_action_data2 <- sim_action_data
sim_action_data2$locked_in <- sim_action_data2$name == "S1_project"
s2 <- ppp_plot_spp_solution(sim_project_data, sim_action_data2,
                            sim_species_data, 300, "name", "success",
                            "name", "cost", "name", "weight",
                            locked_in_column_name = "locked_in")

# print solution
print(s2)

# plot solution
ppp_plot_spp_solution(sim_project_data, sim_action_data2, sim_species_data,
                      s2, "name", "success", "name", "cost", "name",
                      "weight")

# find a solution that meets a budget of 300 and does not allocate
# funding for the "S2_project" project. For instance, species "S2"
# might have very little cultural or economic importance. Broadly speaking,
# though, it is better to "lock in" "important" species rather than
# "lock out" unimportant species.
sim_action_data3 <- sim_action_data
sim_action_data3$locked_out <- sim_action_data3$name == "S2_project"
s3 <- ppp_plot_spp_solution(sim_project_data, sim_action_data3,
                            sim_species_data, 300, "name", "success",
                            "name", "cost", "name", "weight",
                            locked_out_column_name = "locked_out")

# print solution
print(s3)

# plot solution
ppp_plot_spp_solution(sim_project_data, sim_action_data3, sim_tree, s3,
                      "name", "success", "name", "cost")
}
}
\references{
Joseph LN, Maloney RF & Possingham HP (2009) Optimal allocation of
resources among threatened species: A project prioritization protocol.
\emph{Conservation Biology}, \strong{23}, 328--338.

Rodrigues AS & Gaston KJ (2002) Optimisation in reserve selection
procedures---why not? \emph{Biological Conservation}, \strong{107}: 123-129.

Underhill LG (1994) Optimal and suboptimal reserve selection
algorithms. \emph{Biological Conservation}, \strong{70}: 85--87.
}
\seealso{
For other methods for solving the 'Project Prioritization Protocol'
  problem, see \code{\link{ppp_heuristic_phylo_solution}},
  \code{\link{ppp_manual_phylo_solution}}, and
  \code{\link{ppp_random_phylo_solution}}.
  To visualize the effectiveness of a particular solution, see
  \code{\link{ppp_plot_phylo_solution}}.
}
