% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_gurobi_solution.R
\name{ppp_gurobi_solution}
\alias{ppp_gurobi_solution}
\title{Project prioritization protocol}
\usage{
ppp_gurobi_solution(x, tree, budget, cost_column_name, success_column_name,
  locked_in_column_name = NULL, locked_out_column_name = NULL,
  gap = 1e-06, threads = 1L, time_limit = .Machine$integer.max,
  verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which projects should be locked into the funding scheme. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which projects should be locked out of the funding scheme. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{gap}{\code{numeric} optimality gap. Defaults to 0.000001, so that
solutions will be within 0.0001 % of optimality. No missing values are
permitted.}

\item{threads}{\code{numeric} number of threads for computational processing.
Defaults to \code{1}.}

\item{time_limit}{\code{numeric} maximum number of seconds that should be
spent searching for a solution after formatting the data. Effectively,
defaults to no time limit (but specifically is
\code{.Machine$integer.max}).}

\item{verbose}{\code{logical} should information be printed while solving
the problem? Defaults to \code{FALSE}.}

\item{threads}{}
}
\value{
A \code{\link[tibble]{tibble}} object that contains the data in
  the argument to \code{x}, with an additional column named
  \code{"solution"} that contains \code{logical} (i.e.
  \code{TRUE}/\code{FALSE} values) which indicate if each project is
  selected for funding in the solution. This object also has the
  following attributes:

  \describe{

    \item{\code{objective}}{\code{numeric} objective value associated with
      the solution.}

    \item{\code{runtime}}{\code{numeric} number of seconds that elapsed
      while solving the problem.}

    \item{\code{status}}{\code{character} description of the state of the
      solver (e.g. \code{"OPTIMAL"} indicates that the
      optimal solution was found.}

 }
}
\description{
Prioritize funding for conservation projects. To identify optimal funding
schemes, the Gurobi optimization software suite and the
\pkg{gurobi} package need to be installed.
}
\details{
TODO
}
