% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_heuristic_solution.R
\name{ppp_heuristic_solution}
\alias{ppp_heuristic_solution}
\title{Solve the 'Project Prioritization Protocol' problem using heuristic
algorithms}
\usage{
ppp_heuristic_solution(x, tree, budget, project_column_name,
  cost_column_name, success_column_name, locked_in_column_name = NULL,
  locked_out_column_name = NULL)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which projects should be locked into the funding scheme. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which projects should be locked out of the funding scheme. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}
}
\value{
A \code{\link[tibble]{tibble}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{objective}}{\code{numeric} objective value associated with
     each of the solution(s). This corresponds to the expected amount of
     evolutionary history that is retained by each solution.}

   \item{\code{cost}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{optimal}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal (\code{TRUE}) or not (\code{FALSE}.
     Missing values (\code{NA}) indicate that optimality is unknown
     (i.e. because the method used to produce the solution(s) does not
     provide any bounds on their quality).}

   \item{\code{method}}{\code{character} name of method used to produce the
     solution(s).)}

 }
}
\description{
Prioritize funding for conservation projects using a stingy heuristic
algorithm (Bennett \emph{et al}. 2014; Joseph \emph{et al. 2009}).
\strong{Although this algorithm can deliver solutions that perform
better than random, it is extremely unlikely to identify solutions that
are optimal.}.
}
\details{
Briefly, this algorithm works by starting off with all
conservation projects selected for funding and then begins iteratively
defunding (removing) projects until the budget is met (Bennett \emph{et al}.
2014; Joseph \emph{et al. 2009}). In a given iteration,
each project is evaluated in terms of the amount of evolutionary
history that is expected to be lost per unit cost when the project is not
funded. Although this algorithm may identify funding schemes that perform
better than random, it is worth noting that this algorithm is extremely
unlikely to identify optimal solutions. These calculations can be expressed
mathematically using the follow definitions.

To calculate the relative benefit for funding a given project (\eqn{j})
among a set of projects (\eqn{J}), let the expected amount of evolutionary
history that will persist into the future when all the projects are funded
be expressed as \eqn{P(J)}. Also, let the expected amount of evolutionary
history that will persist into the future when all the remaining projects
are funded except for project \eqn{j} be expressed as \eqn{P(J - j)}.
Furthermore, allow the cost for funding project \eqn{j} to be \eqn{C_j}.
Given this, the benefit for funding project \eqn{j} (\eqn{B_j}) in a given
iteration can be expressed as:

\deqn{B_j = \big(P(J) - P(J - j)\big) / C_j}{B_j = (P(J) - P(J - j)) / C_j}

To calculate the expected amount of evolutionary history that will persist
into the future for a given set of funded projects, we will adopt a new set
of definitions to avoid confusion. Let \eqn{I} represent a given set of
funded projects (indexed by \eqn{i}). For example, \eqn{I} could denote all
of the projects in a given iteration (\eqn{P(J)}) or all of the project in a
given iteration except for a specific project (\eqn{P(J - j)}). Next, let
\eqn{P_i} represent the probability of project \eqn{i} being successful if
it is funded. Also, let $S$ represent each species (e.g. species; indexed by
\eqn{s}). To represent the conservation outcome for funding each project,
let $B_{is}$ denote the probability of persistence for the species \eqn{s} if
project \eqn{i} is funded and project $i$ is used to conserve that species.

The probability that each species will go extinct (\eqn{E_s}) when a given
set of projects are funded (\eqn{I}) can then be  expressed as as:

\deqn{E_s = 1 - \text{max}\(B_{1s}, \text{\ldots}, B_{Is}\)}{E_s = 1 - max(B_{1s}, ..., B_{Is})}

To account for the phylogenetic contributions of funding a project,
consider a phylogenetic tree that contains species \eqn{s \in S}{s in S} and
contains branches with known lengths. To describe the tree using mathematical
notation, let \eqn{B} represent the branches (indexed by \eqn{b}) with
lengths \eqn{L_b} and let \eqn{T_{bs}} indicate which species
\eqn{s \in S}{s in S} are associated with which
phylogenetic branches \eqn{b \in B}{b in B} (using zeros and ones).

The amount of evolutionary history that is expected to persist when a
given set of projects are funded can be expressed as:

\deqn{P\(I\) = \sum_{b = 0}^{B} L_b \times \\ \(1 - \prod_{s = 0}^{S} ifelse\(T_{bs} == 1, E_s, 1\)\)}{P(I) = sum_{b = 0}^{B} L_b (1 - prod_{s = 0}^{S} ifelse(T_{bs} == 1, E_s, 1))}
}
\references{
Bennett, J. R., Elliott, G., Mellish, B., Joseph, L. N., Tulloch, A. I.,
Probert, W. J., ... & Maloney, R. (2014). Balancing phylogenetic diversity
and species numbers in conservation prioritization, using a case study of
threatened species in New Zealand. \emph{Biological Conservation},
\strong{174}, 47-54.

Joseph, L. N., Maloney, R. F., & Possingham, H. P. (2009). Optimal
allocation of resources among threatened species: a project prioritization
protocol. \emph{Conservation biology}, \strong{23}, 328--338.
}
\seealso{
For other methods for generating solutions for the 'Project
  Prioritization Protocol' problem,see \code{\link{ppp_heuristic_solution}}
  \code{\link{ppp_gurobi_solution}}, and \code{\link{ppp_random_solution}}.
  To visualize the effectiveness of a particular solution, see
  \code{\link{ppp_plot}}.
}
