% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_manual_phylo_solution.R
\name{ppp_manual_phylo_solution}
\alias{ppp_manual_phylo_solution}
\title{Manually prioritize conservation projects with phylogenetic data}
\usage{
ppp_manual_phylo_solution(x, y, tree, solution, project_column_name,
  success_column_name, action_column_name, cost_column_name)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain data that correspond to
each project. This object should contain data that denote (i)
the name of each project (specified in the argument to
\code{project_column_name}), (ii) the
probability that each project will succeed if all of its actions are funded
(specified in the argument to \code{success_column_name}), (iii)
the enhanced probability that each species will persist if it
is funded (specified as the tip labels in the argument to \code{tree}),
and (iv) and which actions are associated with which projects
(specified in the action names in the argument to \code{y}).
To account for the combined benefits of multiple actions (e.g. baiting
and trapping different invasive species in the same area), additional
projects should be created that indicate the combined cost and
corresponding species' persistence probabilities. Furthermore, this object
must have a baseline project, with a zero cost, that represents the
probability that each species will persist if no other conservation
project is funded.}

\item{y}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing the action data. Here, each row should correspond to a
different action and columns should contain data that correspond to
each action. This object should contain data that denote (i)
the name of each action (specified in the argument to
\code{action_column_name}), (ii) the cost of each action (specified in the
argument to \code{cost_column_name}). If certain actions should be locked
in or out of the solution, then this object should also contain data
that denote (iii) which actions should be locked in (specified using
the argument to \code{locked_in_column_name} if relevant) and (iv) which
actions should be locked out (specified using the argument to
\code{locked_out_column_name} if relevant).}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{solution}{\code{\link[tibble]{tibble}} object containing the solution
data. Here, each row corresponds to a different solution. This object
should contain a column for each action---with the column names matching
the project names---and the column should contain \code{logical} values
that indicate if the action is funded (\code{TRUE}) or not (\code{FALSE}).
Although additional columns can also be included, they will be ignored.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. This argument corresponds to the
argument to \code{x}. Note that the project names must not contain any
duplicates or missing values.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed. This argument
corresponds to the argument to \code{x}. This column must have
\code{numeric} values which lay between zero and one. No missing values
are permitted.}

\item{action_column_name}{\code{character} name of column that contains
the name for each conservation action. This argument corresponds to the
argument to \code{y}. Note that the project names must not contain any
duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each action. This argument corresponds
to the argument to \code{y}. This column must have \code{numeric} values
which are equal to or greater than zero. No missing values are permitted.}
}
\value{
A \code{\link[tibble]{tibble}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{"method"}}{\code{character} name of method used to produce the
     solution(s).)}

   \item{\code{"budget"}}{\code{numeric} budget used for generating each of
     the of the solution(s).}

   \item{\code{"epd"}}{\code{numeric} expected phylogenetic diversity
     associated with each of the solutions. This
     corresponds to the expected amount of evolutionary history that is
     retained by each solution (Faith 2008). If no phylogenetic tree
     was specified, then this column will contain missing values
     (i.e. \code{NA_real_})}.

   \item{\code{"er"}}{\code{numeric} expected species richness. This is
     the sum of the species' persistence probabilities.}

   \item{\code{"cost"}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{"optimal"}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal (\code{TRUE}) or not (\code{FALSE}.
     Missing values (\code{NA}) indicate that optimality is unknown
     (i.e. because the method used to produce the solution(s) does not
     provide any bounds on their quality).}

 }
}
\description{
Manually specify funding schemes for conservation projects using the
'Project Prioritization Protocol' (Joseph, Maloney & Possingham 2009), and
evaluate their effectiveness using 'expected phylogenetic diversity'
(Faith 2008).
}
\examples{
# set seed for reproducibility
set.seed(500)

# load built-in data
data(sim_project_data, sim_action_data, sim_tree)

# load packages to help with plotting
library(ggplot2)

# print simulated project data
print(sim_project_data)

# print simulated action data
print(sim_action_data)

# print simulated phylogenetic tree data
print(sim_tree)

# plot the simulated phylogeny
plot(sim_tree, main = "simulated phylogeny")

# now we will create three solutions
# first, we will initialize a data.frame with no actions funded
solutions <- as.data.frame(matrix(FALSE, ncol = nrow(sim_action_data),
                           nrow = 3))
names(solutions) <- sim_action_data$name

# the first solution will have no actions funded except for the base line
# actions, so we will make the first value in that column TRUE
solutions$baseline_action[1] <- TRUE

# the second solution will have all actions funded, so we will set those
# values to TRUE
solutions[2, ] <- TRUE

# the third solution will have four randomly selected actions funded
solutions[3, sample.int(nrow(sim_action_data), 4)] <- TRUE

# now we can evaluate the solutions
s1 <- ppp_manual_phylo_solution(sim_project_data, sim_action_data, sim_tree,
                                solutions, "name", "cost", "success")

# print the output
print(s1)

# visualize the effectiveness of the different solutions
ppp_plot(sim_project_data, sim_action_data, sim_tree, s1, "name", "cost",
         "success", n = 1) + ggtitle("solution 1")
ppp_plot(sim_project_data, sim_action_data, sim_tree, s1, "name", "cost",
         "success", n = 2) + ggtitle("solution 2")
ppp_plot(sim_project_data, sim_action_data, sim_tree, s1, "name", "cost",
         "success", n = 2) + ggtitle("solution 3")
}
\references{
Faith DP (2008) Threatened species and the potential loss of
phylogenetic diversity: conservation scenarios based on estimated extinction
probabilities and phylogenetic risk analysis. \emph{Conservation Biology},
\strong{22}: 1461--1470.

Joseph LN, Maloney RF & Possingham HP (2009) Optimal allocation of
resources among threatened species: A project prioritization protocol.
\emph{Conservation Biology}, \strong{23}, 328--338.
}
\seealso{
For other methods for generating solutions for the 'Project
  Prioritization Protocol' problem using phylogenetic data, see
  \code{\link{ppp_exact_phylo_solution}},
  \code{\link{ppp_heuristic_phylo_solution}}, and
  \code{\link{ppp_random_phylo_solution}}. To visualize the effectiveness of
  a particular solution, see \code{\link{ppp_phylo_plot}}.
}
