% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_manual_solution.R
\name{ppp_manual_solution}
\alias{ppp_manual_solution}
\title{Solve the 'Project Prioritization Protocol' problem by manually specifying
a solution}
\usage{
ppp_manual_solution(x, tree, budget, solution, project_column_name,
  cost_column_name, success_column_name, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{verbose}{\code{logical} should information be printed while solving
the problem? No missing values are permitted. Defaults to \code{FALSE}.}
}
\description{
Manually specify funding schemes for conservation projects and examine
their effectiveness.
}
\details{
TODO
}
\seealso{
The following methods can be used to extract, inspect, and
  visualize the solution(s): TODO
}
