% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ppp_objective_value}
\alias{ppp_objective_value}
\title{'Project Prioritization Protocol' objective value}
\usage{
ppp_objective_value(x, tree, project_column_name, success_column_name,
  solution)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain data that correspond to
each project. This object should contain data that denote (i)
the name of each project (specified in the argument to
\code{project_column_name}), (ii) the cost of each project
(specified in the argument to \code{cost_column_name}), (iii) the
probability that each project will succeed if it is funded
(specified in the argument to \code{success}),
and (iv) the enhanced probability that each species will persist if it
is funded (specified as the tip labels in the argument to \code{tree}).
To account for the combined benefits of multiple actions (e.g. baiting
and trapping different invasive species in the same area), additional
projects should be created that indicate the combined cost and
corresponding species' persistence probabilities. Furthermore, this object
must have a baseline project, with a zero cost, that represents the
probability that each species will persist if no other conservation
project is funded.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{solution}{\code{\link[tibble]{tibble}} object containing the solution
data. Here, each row corresponds to a different solution. This object
should contain a column for each project---with the column name matching
the project name---and the column should contain \code{logical} values
that indicate if the project is funded (\code{TRUE}) or not (\code{FALSE}).
Although additional columns can also be included, they will be ignored.}
}
\value{
\code{numeric} vector with an objective value for each solution
 in the argument to \code{solutions}.
}
\description{
Calculate the objective value for a set of solutions to a Project
Prioritization Protocol problem.
}
\keyword{internal}
