% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_plot.R
\name{ppp_plot}
\alias{ppp_plot}
\title{Plot a 'Project Prioritization Protocol' solution}
\usage{
ppp_plot(x, tree, project_column_name, success_column_name, solution,
  n = 1)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{solution}{\code{\link[tibble]{tibble}} object containing the solution
data. Here, each row corresponds to a different solution. This object
should contain a column for each project---with the column name matching
the project name---and the column should contain \code{logical} values
that indicate if the project is funded (\code{TRUE}) or not (\code{FALSE}).
Although additional columns can also be included, they will be ignored.}
}
\description{
Create a plot displaying a phylogenetic tree (phylogram) to visualize a
solution to the 'Project Prioritization Protocol' problem. In this plot, each
phylogenetic branch is colored according to probability
that it is expected to persist into the future. Additionally, species that
benefit from at least a single funded project with a non-zero cost are
denoted with large labels.
}
\details{
This function requires the \pkg{ggtree} package (). Since the
  \pkg{ggtree} package is distributed exclusively through
  \href{https://bioconductor.org}{Bioconductor}, please execute the
  following commands to install it:
  \code{source("https://bioconductor.org/biocLite.R");biocLite("ggtree")}.
  If the installation process fails, please consult the \pkg{ggtree}
  package's
  \href{online documentation}{https://bioconductor.org/packages/release/bioc/html/ggtree.html}.
  For more information on customizing the plot, please refer to the package's
  vignettes.
}
\references{
Yu, G., Smith, D. K., Zhu, H., Guan, Y., & Lam, T. T. Y. (2017). ggtree: An
R package for visualization and annotation of phylogenetic trees with their
covariates and other associated data. \emph{Methods in Ecology and
Evolution}, \strong{8}, 28--36.
}
\seealso{
To generate solutions for the 'Project
  Prioritization Protocol' problem, see \code{\link{ppp_heuristic_solution}}
  \code{\link{ppp_gurobi_solution}}, \code{\link{ppp_manual_solution}},
  or \code{\link{ppp_random_solution}}.
}
