% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_plot_phylo_solution.R
\name{ppp_plot_phylo_solution}
\alias{ppp_plot_phylo_solution}
\title{Plot a 'Project Prioritization Protocol' solution with phylogenetic data}
\usage{
ppp_plot_phylo_solution(x, y, tree, solution, project_column_name,
  success_column_name, action_column_name, cost_column_name, n = 1L,
  symbol_hjust = 0.007)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain data that correspond to
each project. This object should contain data that denote (i)
the name of each project (specified in the argument to
\code{project_column_name}), (ii) the
probability that each project will succeed if all of its actions are funded
(specified in the argument to \code{success_column_name}), (iii)
the enhanced probability that each species will persist if it
is funded,
and (iv) and which actions are associated with which projects
(specified in the action names in the argument to \code{y}).
To account for the combined benefits of multiple actions (e.g. baiting
and trapping different invasive species in the same area), additional
projects should be created that indicate the combined cost and
corresponding species' persistence probabilities. Furthermore, this object
must have a baseline project, with a zero cost, that represents the
probability that each species will persist if no other conservation
project is funded.}

\item{y}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing the action data. Here, each row should correspond to a
different action and columns should contain data that correspond to
each action. This object should contain data that denote (i)
the name of each action (specified in the argument to
\code{action_column_name}), (ii) the cost of each action (specified in the
argument to \code{cost_column_name}). If certain actions should be locked
in or out of the solution, then this object should also contain data
that denote (iii) which actions should be locked in (specified using
the argument to \code{locked_in_column_name} if relevant) and (iv) which
actions should be locked out (specified using the argument to
\code{locked_out_column_name} if relevant).}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{solution}{\code{\link[tibble]{tibble}} object containing the solution
data. Here, each row corresponds to a different solution. This object
should contain a column for each action---with the column names matching
the action names (i.e. the argument to \code{y}---and the columns should
contain \code{logical} values that indicate if the action is funded
(\code{TRUE}) or not (\code{FALSE}). Although additional columns can also
be included, they will be ignored.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. This argument corresponds to the
argument to \code{x}. Note that the project names must not contain any
duplicates or missing values.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed. This argument
corresponds to the argument to \code{x}. This column must have
\code{numeric} values which lay between zero and one. No missing values
are permitted.}

\item{action_column_name}{\code{character} name of column that contains
the name for each conservation action. This argument corresponds to the
argument to \code{y}. Note that the project names must not contain any
duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each action. This argument corresponds
to the argument to \code{y}. This column must have \code{numeric} values
which are equal to or greater than zero. No missing values are permitted.}

\item{n}{\code{integer} solution to display. This should argument
correspond to a row number in the argument to \code{solution}. Defaults
to \code{1} such that the solution in the first row of the argument
to \code{solution} is plotted.}

\item{symbol_hjust}{\code{numeric} horizontal adjustment parameter to
manually align the asterisks and dashes in the plot. Defaults to
\code{0.007}. Increasing this parameter will shift the symbols further
left towards the species labels.}
}
\description{
Create a plot displaying a phylogenetic tree (phylogram) to visualize a
solution to the 'Project Prioritization Protocol' problem
(Joseph, Maloney & Possingham 2009) with phlyogenetic data.
}
\details{
This function requires the \pkg{ggtree} (Yu \emph{et al.} 2017).
  Since this package is distributed exclusively
  through \href{https://bioconductor.org}{Bioconductor}, and is not
  available on the
  \href{https://cran.r-project.org/}{Comprehensive R Archive Network},
  please execute the following command to install it:
  \code{source("https://bioconductor.org/biocLite.R");biocLite("ggtree")}.
  If the installation process fails, please consult the package's \href{https://bioconductor.org/packages/release/bioc/html/ggtree.html}{online documentation}.

  Here, each phylogenetic branch is colored according to probability
  that it is expected to persist into the future (based on Faith 2008).
  Additionally, species that directly benefit from at least a single
  completely funded project with a non-zero cost are denoted with an
  asterisk symbol. Species that indirectly benefit from funded
  projects---because they are associated with partially funded projects that
  have non-zero costs and share actions with at least one funded
  project---are denoted with an open circle symbols.
}
\examples{
# set seed for reproducibility
set.seed(500)

# load built-in data
data(sim_project_data, sim_action_data, sim_tree)

# print simulated project data
print(sim_project_data)

# print simulated action data
print(sim_action_data)

# print simulated phylogenetic tree data
print(sim_tree)

# plot the simulated phylogeny
plot(sim_tree, main = "simulated phylogeny")

# create random some solutions with a budget of 700
s1 <- ppp_random_phylo_solution(sim_project_data, sim_action_data, sim_tree,
                                700, "name", "success", "name", "cost",
                                number_solutions = 10)

# print output
print(s1)

# plot the first solution
ppp_plot_phylo_solution(sim_project_data, sim_action_data, sim_tree, s1,
                        "name", "success", "name", "cost")

# plot the second solution
ppp_plot_phylo_solution(sim_project_data, sim_action_data, sim_tree, s1,
                        "name", "success", "name", "cost", n = 2)

# since this function returns a ggplot2 plot object, we can customize the
# appearance of the plot using standard ggplot2 commands!
# for example, we can add a title
ppp_plot_phylo_solution(sim_project_data, sim_action_data, sim_tree, s1,
                        "name", "success", "name", "cost") +
ggtitle("solution")

# we could also also set the minimum and maximum values in the color ramp to
# correspond to those in the data, rather than being capped at 0 and 1
ppp_plot_phylo_solution(sim_project_data, sim_action_data, sim_tree, s1,
                        "name", "success", "name", "cost") +
scale_color_gradientn(name = "Probability of\\npersistence",
                      colors = viridisLite::inferno(150, begin = 0,
                                                    end = 0.9,
                                                    direction = -1)) +
ggtitle("solution")

# we could also change the color ramp
ppp_plot_phylo_solution(sim_project_data, sim_action_data, sim_tree, s1,
                        "name", "success", "name", "cost") +
scale_color_gradient(name = "Probability of\\npersistence",
                     low = "red", high = "black") +
ggtitle("solution")

# we could even hide the legend if desired
ppp_plot_phylo_solution(sim_project_data, sim_action_data, sim_tree, s1,
                        "name", "success", "name", "cost") +
scale_color_gradient(name = "Probability of\\npersistence",
                     low = "red", high = "black") +
theme(legend.position = "hide") +
ggtitle("solution")
}
\references{
Faith DP (2008) Threatened species and the potential loss of
phylogenetic diversity: conservation scenarios based on estimated extinction
probabilities and phylogenetic risk analysis. \emph{Conservation Biology},
\strong{22}: 1461--1470.

Joseph LN, Maloney RF & Possingham HP (2009) Optimal allocation of
resources among threatened species: A project prioritization protocol.
\emph{Conservation Biology}, \strong{23}, 328--338.

Yu G, Smith DK, Zhu H, Guan Y, & Lam TTY (2017) ggtree: an
R package for visualization and annotation of phylogenetic trees with their
covariates and other associated data. \emph{Methods in Ecology and
Evolution}, \strong{8}: 28--36.
}
\seealso{
To generate solutions for the 'Project
  Prioritization Protocol' problem, see
  \code{\link{ppp_heuristic_phylo_solution}}
  \code{\link{ppp_exact_phylo_solution}},
  \code{\link{ppp_manual_phylo_solution}},
  or \code{\link{ppp_random_phylo_solution}}.
}
