% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_random_phylo_solution.R
\name{ppp_random_phylo_solution}
\alias{ppp_random_phylo_solution}
\title{Randomly prioritize conservation projects under expected phylogenetic
diversity}
\usage{
ppp_random_phylo_solution(x, y, tree, budget, project_column_name,
  success_column_name, action_column_name, cost_column_name,
  locked_in_column_name = NULL, locked_out_column_name = NULL,
  number_solutions = 1L)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain data that correspond to
each project. This object should contain data that denote (i)
the name of each project (specified in the argument to
\code{project_column_name}), (ii) the
probability that each project will succeed if all of its actions are funded
(specified in the argument to \code{success_column_name}), (iii)
the enhanced probability that each species will persist if it
is funded (specified as the tip labels in the argument to \code{tree}),
and (iv) and which actions are associated with which projects
(specified in the action names in the argument to \code{y}).
To account for the combined benefits of multiple actions (e.g. baiting
and trapping different invasive species in the same area), additional
projects should be created that indicate the combined cost and
corresponding species' persistence probabilities. Furthermore, this object
must have a baseline project, with a zero cost, that represents the
probability that each species will persist if no other conservation
project is funded.}

\item{y}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing the action data. Here, each row should correspond to a
different action and columns should contain data that correspond to
each action. This object should contain data that denote (i)
the name of each action (specified in the argument to
\code{action_column_name}), (ii) the cost of each action (specified in the
argument to \code{cost_column_name}). If certain actions should be locked
in or out of the solution, then this object should also contain data
that denote (iii) which actions should be locked in (specified using
the argument to \code{locked_in_column_name} if relevant) and (iv) which
actions should be locked out (specified using the argument to
\code{locked_out_column_name} if relevant).}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. This argument corresponds to the
argument to \code{x}. Note that the project names must not contain any
duplicates or missing values.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed. This argument
corresponds to the argument to \code{x}. This column must have
\code{numeric} values which lay between zero and one. No missing values
are permitted.}

\item{action_column_name}{\code{character} name of column that contains
the name for each conservation action. This argument corresponds to the
argument to \code{y}. Note that the project names must not contain any
duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each action. This argument corresponds
to the argument to \code{y}. This column must have \code{numeric} values
which are equal to or greater than zero. No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which actions should be locked into the funding scheme.
This argument corresponds to the argument to \code{y}. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which actions should be locked out of the funding scheme.
This argument corresponds to the argument to \code{y}. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{number_solutions}{\code{numeric} number of solutions to return. If
the argument is greater than \code{1}, then the output will contain the
set number of solutions that are closest to optimality. No missing values
are permitted. Defaults to \code{1}.}
}
\description{
Generate random solutions for the 'Project Prioritization Protocol'
problem (Joseph, Maloney & Possingham 2009).
Although conservation projects should, ideally, not be funded based on random
allocations, it can be useful to compare the effectiveness of solutions to
random decisions in order to evaluate their effectiveness.
\strong{When informing conservation actions, it is strongly recommended to
use the \code{\link{ppp_exact_phylo_solution}} method because it can identify
optimal funding schemes with a guarantee.}
}
\details{
The random solutions are generated using the following algorithm.
 Firstly, all projects are initially selected for funding (excepting actions
 which are locked out). Secondly, an action is randomly selected and
 defunded. Thirdly, the second step is
 repeated until the total cost of the remaining projects that are selected
 for funding is within the budget. Note that actions that have zero cost
 are never deselected for funding, and are always included in the solutions.
 Additionally, actions that are locked in are never deselected for funding.
}
\examples{
# set seed for reproducibility
set.seed(500)

# load built-in data
data(sim_project_data, sim_action_data, sim_tree)

# print simulated project data
print(sim_project_data)

# print simulated action data
print(sim_action_data)

# print simulated phylogenetic tree data
print(sim_tree)

# plot the simulated phylogeny
plot(sim_tree, main = "simulated phylogeny")

# generate 10 random solutions that meet a budget of 300
s1 <- ppp_random_phylo_solution(sim_project_data, sim_action_data, sim_tree,
                                300, "name", "success", "name", "cost",
                                number_solutions = 10)

# print solutions
print(s1)

# plot first solution
ppp_plot_phylo_solution(sim_project_data, sim_action_data, sim_tree, s1,
                        "name", "success", "name", "cost", n = 1)

# view histogram of their expected phylogenetic diversity
hist(s1$obj, xlab = "expected phylogenetic diversity")

# view histogram of their costs
hist(s1$cost, xlab = "solution cost")
}
\references{
Faith DP (2008) Threatened species and the potential loss of
phylogenetic diversity: conservation scenarios based on estimated extinction
probabilities and phylogenetic risk analysis. \emph{Conservation Biology},
\strong{22}: 1461--1470.

Joseph LN, Maloney RF & Possingham HP (2009) Optimal allocation of
resources among threatened species: A project prioritization protocol.
\emph{Conservation Biology}, \strong{23}, 328--338.
}
\seealso{
For other methods for generating solutions for the 'Project
  Prioritization Protocol' problem using phylogenetic data, see
  \code{\link{ppp_heuristic_phylo_solution}}
  \code{\link{ppp_exact_phylo_solution}}, and
  \code{\link{ppp_manual_phylo_solution}}.
  To visualize the effectiveness of a particular solution, see
  \code{\link{ppp_plot_phylo_solution}}.
}
