% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_random_solution.R
\name{ppp_random_solution}
\alias{ppp_random_solution}
\title{Random solutions for the 'Project Prioritization Protocol' problem}
\usage{
ppp_random_solution(x, tree, budget, project_column_name, cost_column_name,
  success_column_name, locked_in_column_name = NULL,
  locked_out_column_name = NULL, number_solutions = 1L)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which projects should be locked into the funding scheme. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which projects should be locked out of the funding scheme. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{number_solutions}{\code{numeric} number of solutions to return. If
the argument is greater than \code{1}, then the output will contain the
set number of solutions that are closest to optimality. No missing values
are permitted. Defaults to \code{1}.}
}
\description{
Generate random solutions for the 'Project Prioritization Protocol' problem.
Although conservation projects should, ideally, not be funded based on random
allocations, it can be useful to compare the effectiveness of solutions to
random decision in order to evaluate their effectiveness.
\strong{When informing conservation actions, it is strongly recommended to
use the \code{\link{ppp_exact_solution}} method because it can identify
optimal funding schemes with a guarantee.}
}
\details{
The random solutions are generated using the following algorithm.
 All projects are initially selected for funding. Next,
 a project is randomly deselected for funding (removed), and this step
 is repeated until the cost of the remaining projects is within the
 budget. Note that projects that have zero cost are never deselected for
 funding, and are always included in the solutions. The different solutions
 are then evaluated according to the 'expected phylogenetic diversity'
 metric (Faith 2008).
}
\examples{
# load built-in data
data(sim_project_data, sim_tree)

# print simulated project data set
print(sim_project_data)

# print simulated phylogenetic tree data set
print(sim_tree)

# plot the simulated phylogeny
plot(sim_tree, main = "simulated phylogeny")

# generate 10 random solutions that meet a budget of 300
s1 <- ppp_random_solution(sim_project_data, sim_tree, 300,
                          "name", "cost", "success", number_solutions = 10)

# print solutions
print(s1)

# plot first solution
ppp_plot(sim_project_data, sim_tree, s1, "name", "cost", "success", n = 1)

# view histogram of their objective values
hist(s1$objective, xlab = "solution objective")

# view histogram of their costs
hist(s1$cost, xlab = "solution cost")

}
\references{
Faith DP (2008) Threatened species and the potential loss of
phylogenetic diversity: conservation scenarios based on estimated extinction
probabilities and phylogenetic risk analysis. \emph{Conservation Biology},
\strong{22}: 1461--1470.
}
\seealso{
For other methods for generating solutions for the 'Project
  Prioritization Protocol' problem, see \code{\link{ppp_heuristic_solution}}
  \code{\link{ppp_exact_solution}}, and \code{\link{ppp_manual_solution}}.
  To visualize the effectiveness of a particular solution, see
  \code{\link{ppp_plot}}.
}
