% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_results_class.R
\name{ppp_results_class}
\alias{ppp_results_class}
\title{'Project Prioritization Protocol' solution class}
\description{
This manual page documents information about the
\code{\link[tibble]{tbl_df}} object output from the solver functions.
}
\seealso{
The following methods can be used to extract, inspect, and
  visualize the solution(s): TODO
}
\keyword{internal}
