% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\alias{sim_tree}
\alias{sim_project_data}
\title{Simulated project and phylogenetic data}
\format{\describe{
  \item{sim_project_data}{\code{\link[base]{data.frame}} object}
  \item{sim_tree}{\code{\link[ape]{phylo}} object.}
}}
\usage{
data(sim_project_data)

sim_project_data

sim_tree
}
\description{
Simulated data for prioritizing conservations projects.
}
\details{
The data set contains the following objects:

\describe{

  \item{\code{sim_project_data}}{A \code{\link[base]{data.frame}} containing
    data for six simulated conservation projects. Each row corresponds to a
    different project and each column contains information about the
    projects. This table contains the following columns:

    \describe{

      \item{\code{"name"}}{\code{character} name for each project.}

      \item{\code{"cost"}}{\code{numeric} cost for each project.}

      \item{\code{"success"}}{\code{numeric} probability of each project
        succeeding if it is funded.}

      \item{\code{"locked_in"}}{\code{logical} indicating if certain
        projects should be locked into the solution.}

      \item{\code{"locked_out"}}{\code{logical} indicating if certain
        projects should be locked out of the solution.}

      \item{\code{"S1"} ... \code{"S5"}}{\code{numeric} columns for each
      species (i.e. \code{"S1"}, \code{"S2"}, \code{"S3"}, \code{"S4"},
      \code{"S5"}, indicating the enhanced probability that each
        species will survive if it funded.}

    }}

  \item{\code{sim_tree}}{A \code{\link[ape]{phylo}} phylogenetic tree
    that describes the evolutionary relationships between five simulated
    species. These species are named: \code{"S1"}, \code{"S2"}, \code{"S3"},
    \code{"S4"}, and \code{"S5"}.}

}
}
\examples{
# load data
data(sim_project_data, sim_tree)

# print project data
print(sim_project_data)
# print phylogenetic tree
print(sim_tree)

# plot phylogenetic tree
plot(sim_tree)

}
\keyword{datasets}
