% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_exact_solution.R
\name{ppp_exact_solution}
\alias{ppp_exact_solution}
\title{Solve the 'Project Prioritization Protocol' problem using exact algorithms}
\usage{
ppp_exact_solution(x, tree, budget, project_column_name, cost_column_name,
  success_column_name, locked_in_column_name = NULL,
  locked_out_column_name = NULL, gap = 1e-06, threads = 1L,
  number_solutions = 1L, time_limit = .Machine$integer.max,
  number_approx_points = 300, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which projects should be locked into the funding scheme. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which projects should be locked out of the funding scheme. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{gap}{\code{numeric} optimality gap. This gap should be expressed as
a proportion. For example, to find a solution that is within 10 % of
optimality, then \code{0.1} should be supplied. No missing values are
permitted. Defaults to \code{0}, so that the optimal solution will be
returned.}

\item{threads}{\code{numeric} number of threads for computational processing.
No missing values are permitted. Defaults to \code{1}.}

\item{number_solutions}{\code{numeric} number of solutions to return. If
the argument is greater than \code{1}, then the output will contain the
set number of solutions that are closest to optimality. No missing values
are permitted. Defaults to \code{1}.}

\item{time_limit}{\code{numeric} maximum number of seconds that should be
spent searching for a solution after formatting the data. Effectively,
defaults to no time limit (but specifically is
\code{.Machine$integer.max}). No missing values are permitted.}

\item{number_approx_points}{\code{numeric} number of points to use for
approximating the probability that branches will go extent. Larger
values increase the precision of these calculations.
No missing values are permitted. Defaults to \code{300}.}

\item{verbose}{\code{logical} should information be printed while solving
the problem? No missing values are permitted. Defaults to \code{FALSE}.}
}
\value{
A \code{\link[tibble]{tibble}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{"objective"}}{\code{numeric} objective value associated with
     each of the solution(s). This corresponds to the expected amount of
     evolutionary history that is retained by each solution.}

   \item{\code{"budget"}}{\code{numeric} budget used for generating each of
     the of the solution(s).}

   \item{\code{"cost"}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{"optimal"}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal (\code{TRUE}) or not (\code{FALSE}.
     Missing values (\code{NA}) indicate that optimality is unknown
     (i.e. because the method used to produce the solution(s) does not
     provide any bounds on their quality).}

   \item{\code{"method"}}{\code{character} name of method used to produce the
     solution(s).)}

 }
}
\description{
Prioritize funding for conservation projects using exact algorithms. Unlike
other methods for generating prioritizations, this method can identify
solutions that are guaranteed to be optimal (or within a pre-specified
optimality gap; see Underhill 1994; Rodrigues & Gaston 2002).
\strong{As a consequence, it is strongly recommended to use
this method for developing project prioritizations.}
}
\details{
This function works by formulating the 'Project Prioritization
  Protocol' as a mixed integer programming problem (MIP) and solving it
  using the
  \href{https://www.gurobi.com}{Gurobi optimization software suite}.
  Specifically, the problem aims to maximize
  the amount of evolutionary history that is expected to
  persist (i.e. the 'expected phylogenetic diversity' metric; Faith 2008).
  This is achieved by funding different conservation projects, with
  known costs, that have a known effect on species' survival. Please refer
  to the package vignette for the complete formulation of this problem.

  Although \href{https://www.gurobi.com}{Gurobi} is a commercial software,
  academics can obtain a \href{https://user.gurobi.com/download/licenses/free-academic}{special license for no cost}.
   After downloading and installing the
  href{https://www.gurobi.com}{Gurobi} software suite, the
  \pkg{gurobi} package will also need to be installed (see instructions for
  \href{http://www.gurobi.com/documentation/8.1/quickstart_linux/software_installation_guid.html}{Linux},
  \href{http://www.gurobi.com/documentation/8.1/quickstart_mac/software_installation_guid.html}{Mac OSX}, and
  \href{http://www.gurobi.com/documentation/8.1/quickstart_windows/software_installation_guid.html}{Windows} operating systems).
Finally, the \pkg{gurobi} package will also need to be installed (see
instructions for
 \href{http://www.gurobi.com/documentation/8.1/quickstart_linux/r_installing_the_r_package.html}{Linux},
  \href{http://www.gurobi.com/documentation/8.1/quickstart_mac/r_installing_the_r_package.html}{Mac OSX}, and
  \href{http://www.gurobi.com/documentation/8.1/quickstart_windows/r_installing_the_r_package.html}{Windows} operating systems).
}
\examples{
# load built-in data
data(sim_project_data, sim_tree)

# print simulated project data set
print(sim_project_data)

# print simulated phylogenetic tree data set
print(sim_tree)

# plot the simulated phylogeny
plot(sim_tree, main = "simulated phylogeny")
\donttest{
# verify if guorbi package is installed
if (!require(gurobi, quietly = TRUE))
 stop("the gurobi R package is not installed.")

# find a solution that meets a budget of 300
s1 <- ppp_exact_solution(sim_project_data, sim_tree, 300,
                         "name", "cost", "success")

# print solution
print(s1)

# print the names of which projects were funded
print(names(s1)[which(unlist(s1[1, sim_project_data$name]))])

# plot solution
ppp_plot(sim_project_data, sim_tree, s1, "name", "cost", "success")

# find a solution that meets a budget of 300 and allocates
# funding for the "S1_project" project. For instance, species "S1" might
# be an iconic species that has cultural and economic importance.
sim_project_data2 <- sim_project_data
sim_project_data2$locked_in <- sim_project_data2$name == "S1_project"
s2 <- ppp_exact_solution(sim_project_data2, sim_tree, 300,
                         "name", "cost", "success",
                         locked_in_column_name = "locked_in")

# print solution
print(s2)

# plot solution
ppp_plot(sim_project_data2, sim_tree, s2, "name", "cost", "success")

# find a solution that meets a budget of 300 and does not allocate
# funding for the "S2_project" project. For instance, species "S2"
# might have very little cultural or economic importance. Broadly speaking,
# though, it is better to "lock in" "important" species rather than
# "lock out" unimportant species.
sim_project_data3 <- sim_project_data
sim_project_data3$locked_out <- sim_project_data2$name == "S2_project"
s3 <- ppp_exact_solution(sim_project_data3, sim_tree, 300,
                         "name", "cost", "success",
                         locked_out_column_name = "locked_out")

# print solution
print(s3)

# plot solution
ppp_plot(sim_project_data3, sim_tree, s3, "name", "cost", "success")
}
}
\references{
Faith DP (2008) Threatened species and the potential loss of
phylogenetic diversity: conservation scenarios based on estimated extinction
probabilities and phylogenetic risk analysis. \emph{Conservation Biology},
\strong{22}: 1461--1470.

Rodrigues AS & Gaston KJ (2002) Optimisation in reserve selection
procedures---why not? \emph{Biological Conservation}, \strong{107}: 123-129.

Underhill LG (1994) Optimal and suboptimal reserve selection
algorithms. \emph{Biological Conservation}, \strong{70}: 85--87.
}
\seealso{
For other methods for solving the 'Project Prioritization Protocol'
  problem, see \code{\link{ppp_heuristic_solution}},
  \code{\link{ppp_manual_solution}}, and \code{\link{ppp_random_solution}}.
  To visualize the effectiveness of a particular solution, see
  \code{\link{ppp_plot}}.
}
