% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_gurobi_solution.R
\name{ppp_gurobi_solution}
\alias{ppp_gurobi_solution}
\title{Solve the 'Project Prioritization Protocol' problem using Gurobi}
\usage{
ppp_gurobi_solution(x, tree, budget, project_column_name, cost_column_name,
  success_column_name, locked_in_column_name = NULL,
  locked_out_column_name = NULL, gap = 1e-06, threads = 1L,
  number_solutions = 1L, time_limit = .Machine$integer.max,
  number_approx_points = 300, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which projects should be locked into the funding scheme. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which projects should be locked out of the funding scheme. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{gap}{\code{numeric} optimality gap. This gap should be expressed as
a proportion. For example, to find a solution that is within 10 % of
optimality, then \code{0.1} should be supplied. No missing values are
permitted. Defaults to \code{0}, so that the optimal solution will be
returned.}

\item{threads}{\code{numeric} number of threads for computational processing.
No missing values are permitted. Defaults to \code{1}.}

\item{number_solutions}{\code{numeric} number of solutions to return. If
the argument is greater than \code{1}, then the output will contain the
set number of solutions that are closest to optimality. No missing values
are permitted. Defaults to \code{1}.}

\item{time_limit}{\code{numeric} maximum number of seconds that should be
spent searching for a solution after formatting the data. Effectively,
defaults to no time limit (but specifically is
\code{.Machine$integer.max}). No missing values are permitted.}

\item{number_approx_points}{\code{numeric} number of points to use for
approximating the probability that branches will go extent. Larger
values increase the precision of these calculations.
No missing values are permitted. Defaults to \code{300}.}

\item{verbose}{\code{logical} should information be printed while solving
the problem? No missing values are permitted. Defaults to \code{FALSE}.}
}
\description{
Prioritize funding for conservation projects using the Gurobi optimization
software suite. Unlike other methods for generating prioritizations,
this method can identify optimal solutions (or solutions within a
pre-specified optimality gap). \strong{As a consequence, it is strongly
recommended to use this method for developing project prioritizations.}
}
\details{
TODO
}
\seealso{
The following methods can be used to extract, inspect, and
  visualize the solution(s): TODO
}
