% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_gurobi_solution.R
\name{ppp_gurobi_solution}
\alias{ppp_gurobi_solution}
\title{Solve the 'Project Prioritization Protocol' problem using Gurobi}
\usage{
ppp_gurobi_solution(x, tree, budget, cost_column_name, success_column_name,
  locked_in_column_name = NULL, locked_out_column_name = NULL,
  gap = 1e-06, threads = 1L, number_solutions = 1L,
  time_limit = .Machine$integer.max, number_approx_points = 300,
  verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which projects should be locked into the funding scheme. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which projects should be locked out of the funding scheme. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{gap}{\code{numeric} optimality gap. This gap should be expressed as
a proportion. For example, to find a solution that is within 10 % of
optimality, then \code{0.1} should be supplied. No missing values are
permitted. Defaults to \code{0}, so that the optimal solution will be
returned.}

\item{threads}{\code{numeric} number of threads for computational processing.
No missing values are permitted. Defaults to \code{1}.}

\item{number_solutions}{\code{numeric} number of solutions to return. If
the argument is greater than \code{1}, then the output will contain the
set number of solutions that are closest to optimality. No missing values
are permitted. Defaults to \code{1}.}

\item{time_limit}{\code{numeric} maximum number of seconds that should be
spent searching for a solution after formatting the data. Effectively,
defaults to no time limit (but specifically is
\code{.Machine$integer.max}). No missing values are permitted.}

\item{number_approx_points}{\code{numeric} number of points to use for
approximating the probability that branches will go extent. Larger
values increase the precision of these calculations.
No missing values are permitted. Defaults to \code{300}.}

\item{verbose}{\code{logical} should information be printed while solving
the problem? No missing values are permitted. Defaults to \code{FALSE}.}
}
\value{
A \code{ppp_results} class object containing the solutions.
  This is a \code{list} object which contains the following components:

 \describe{

   \item{\code{$call}}{\code{\link[base]{call}} command used to generate the
     solution(s).}

   \item{\code{$project_data}}{\code{\link[tibble]{tbl_df}} object that
     contains the project data used to generate the solution(s).}

   \item{\code{$tree}}{\code{\link[ape]{phylo}} phylogenetic tree used to
     generate the solution(s).}

   \item{\code{$solution}}{\code[\link[tibble]{tbl_df} object that contains a
     column for each solution. Each row corresponds to a different project,
     (following the same order as the data in the argument to \code{x}), and
     each column corresponds to a different solution. Cell values are
     \code{logical} (i.e. \code{TRUE}/\code{FALSE} values) indicating if each
     project is selected for funding in a given solution. The column names
     are formatted where the name \code{"solution_N"} contains the
     the n'th solution. For example, if only one solution is returned
     (as specified in the \code{number_solutions} argument) then the
     output object will only contain a single column called
     \code{"solution_1"}.}

  \item{\code{$statistics}}{\code[\link[tibble]{tbl_df} object that contains
    summary statistics for each solution. Each row corresponds to a
    different solution and each column describes a different property of the
    solution. Specifically, it contains the following columns:

    \describe{

      \item{\code{name}}{\code{character} name of each solution. The solution
        names correspond to the column names in the \code{$solution}
        element.}

      \item{\code{cost}}{\code{numeric} total cost associated with each of
        of the solution(s).}

      \item{\code{objective}}{\code{numeric} objective value associated with
        each of the solution(s). This corresponds to the expected amount of
        evolutionary history that is retained by each solution.}

      \item{\code{optimal}}{\code{logical} indicating if each of the
        solution(s) is known to be optimal or not.)}

    }}

  \item{\code{runtime}}{\code{numeric} number of seconds that elapsed
    while solving the problem.}

  \item{\code{status}}{\code{character} description of the
    solver status (e.g. \code{"OPTIMAL"} indicates that optimal solution(s)
    were found.)}

  \item{\code{method}}{\code{character} name of the method used to produce
    the solutions (e.g. \code{"gurobi"} if solutions were obtained using the
    \code{\link{ppp_gurobi_solution}} function).}

 }
}
\description{
Prioritize funding for conservation projects using the Gurobi optimization
software suite. Unlike other methods for generating prioritizations,
this method can identify optimal solutions (or solutions within a
pre-specified optimality gap). \strong{As a consequence, it is strongly
recommended to use this method for developing project prioritizations.}
}
\details{
TODO
}
\seealso{
The following methods can be used to extract, inspect, and
  visualize components of the
  solution(s): (e.g. extract the solutions, or the summary statistics)
}
