% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_heuristic_solution.R
\name{ppp_heuristic_solution}
\alias{ppp_heuristic_solution}
\title{Solve the 'Project Prioritization Protocol' problem using heuristic
algorithms}
\usage{
ppp_heuristic_solution(x, tree, budget, project_column_name,
  cost_column_name, success_column_name, locked_in_column_name = NULL,
  locked_out_column_name = NULL, number_solutions = 1L)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain data that correspond to
each project. This object should contain data that denote (i)
the name of each project (specified in the argument to
\code{project_column_name}), (ii) the
probability that each project will succeed if all of its actions are funded
(specified in the argument to \code{success_column_name}), (iii)
the enhanced probability that each species will persist if it
is funded (specified as the tip labels in the argument to \code{tree}),
and (iv) and which actions are associated with which projects
(specified in the action names in the argument to \code{y}).
To account for the combined benefits of multiple actions (e.g. baiting
and trapping different invasive species in the same area), additional
projects should be created that indicate the combined cost and
corresponding species' persistence probabilities. Furthermore, this object
must have a baseline project, with a zero cost, that represents the
probability that each species will persist if no other conservation
project is funded.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{locked_in_column_name}{\code{character} name of column that
indicates which projects should be locked into the funding scheme. For
example, it may be desirable to mandate that projects for
iconic species are funded in the prioritization. This column should
contain \code{logical} values, and projects associated with \code{TRUE}
values are locked into the solution. No missing values are permitted.
Defaults to \code{NULL} such that no projects are locked into the
solution.}

\item{locked_out_column_name}{\code{character} name of column that
indicates which projects should be locked out of the funding scheme. For
example, it may be desirable to lock out projects for certain species
that are expected to have little support from the public. This column
should contain \code{logical} values, and projects associated with
\code{TRUE} values are locked out of the solution. No missing values are
permitted. Defaults to \code{NULL} such that no projects are locked out of
the solution.}

\item{number_solutions}{\code{numeric} number of solutions to return. If
the argument is greater than \code{1}, then the output will contain the
set number of solutions that are closest to optimality. No missing values
are permitted. Defaults to \code{1}.}
}
\value{
A \code{\link[tibble]{tibble}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{"method"}}{\code{character} name of method used to produce the
     solution(s).)}

   \item{\code{"budget"}}{\code{numeric} budget used for generating each of
     the of the solution(s).}

   \item{\code{"epd"}}{\code{numeric} expected phylogenetic diversity
     associated with each of the solutions. This
     corresponds to the expected amount of evolutionary history that is
     retained by each solution (Faith 2008). If no phylogenetic tree
     was specified, then this column will contain missing values
     (i.e. \code{NA_real_})}.

   \item{\code{"er"}}{\code{numeric} expected species richness. This is
     the sum of the species' persistence probabilities.}

   \item{\code{"cost"}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{"optimal"}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal (\code{TRUE}) or not (\code{FALSE}.
     Missing values (\code{NA}) indicate that optimality is unknown
     (i.e. because the method used to produce the solution(s) does not
     provide any bounds on their quality).}

 }
}
\description{
Prioritize funding for conservation projects under the 'Project
Prioritization Protocol' (Joseph, Maloney & Possingham 2009) using a stingy
heuristic algorithm (Bennett \emph{et al}. 2014).
\strong{Although this algorithm can deliver solutions that perform
better than random, it is extremely unlikely to identify solutions that
are optimal (Underhill 1994; Rodrigues & Gaston 2002).}
}
\details{
Briefly, this algorithm works by starting off with all
conservation projects selected for funding and then begins iteratively
defunding (removing) projects until the budget is met
(Joseph, Maloney & Possingham 2009; Bennett \emph{et al}. 2014). In a given
iteration,
each project is evaluated in terms of the amount of evolutionary
history that is expected to be lost per unit cost when the project is not
funded (based on the 'expected phylogenetic diversity' metric; Faith 2008),
and the project associated with the lowest utility is defunded.
Although this algorithm may identify funding schemes that perform
better than random, it is worth noting that this algorithm is extremely
unlikely to identify optimal solutions. These calculations can be expressed
mathematically with the following definitions.

To calculate the utility for funding a given project (\eqn{j})
among a set of projects (\eqn{J}), let the expected amount of evolutionary
history that will persist into the future when all the projects are funded
be expressed as \eqn{P(J)}. Also, let the expected amount of evolutionary
history that will persist into the future when all the remaining projects
are funded except for project \eqn{j} be expressed as \eqn{P(J - j)}.
Furthermore, allow the cost for funding project \eqn{j} to be \eqn{C_j}.
Given this, the relative benefit (or utility) for funding project \eqn{j}
(\eqn{U_j}) in a given iteration can be expressed as:

\deqn{U_j = \frac{P(J) - P(J - j)}{C_j}}{U_j = (P(J) - P(J - j)) / C_j}

To calculate the expected amount of evolutionary history that will persist
into the future for a given set of funded projects, we will adopt a new set
of definitions to avoid confusion. Let \eqn{I} represent a given set of
funded projects (indexed by \eqn{i}). For example, \eqn{I} could denote all
of the projects in a given iteration (\eqn{P(J)}) or all of the project in a
given iteration except for a specific project (\eqn{P(J - j)}). Next, let
\eqn{P_i} represent the probability of project \eqn{i} being successful if
it is funded. Also, let \eqn{S} represent each species (e.g. species;
project, let \eqn{B_{is}} denote the probability of persistence for the
species \eqn{s} if project \eqn{i} is funded and project \eqn{i} is used to
conserve that species.

The probability that each species will go extinct (\eqn{E_s}) when a given
set of projects are funded (\eqn{I}) can then be  expressed as as:

\deqn{E_s = 1 - \mathrm{max}(P_1 \times B_{1s}, \ldots, P_I \times B_{Is})}{E_s = 1 - max(P_1 B_{1s}, ..., P_I B_{Is})}

To account for the phylogenetic contributions of funding a project,
consider a phylogenetic tree that contains species \eqn{s \in S}{s in S} and
contains branches with known lengths. To describe the tree using mathematical
notation, let \eqn{B} represent the branches (indexed by \eqn{b}) with
lengths \eqn{L_b} and let \eqn{T_{bs}} indicate which species
\eqn{s \in S}{s in S} are associated with which
phylogenetic branches \eqn{b \in B}{b in B} (using zeros and ones).

The amount of evolutionary history that is expected to persist when a
given set of projects are funded can then be expressed as:

\deqn{P(I) = \sum_{b = 0}^{B} L_b \times \big(1 - \prod_{s = 0}^{S}
  ifelse(T_{bs} == 1, E_s, 1)\big)}{P(I) = sum_{b = 0}^{B} L_b (1 -
  prod_{s = 0}^{S} ifelse(T_{bs} == 1, E_s, 1))}
}
\examples{
# load built-in data
data(sim_project_data, sim_tree)

# print simulated project data set
print(sim_project_data)

# print simulated phylogenetic tree data set
print(sim_tree)

# plot the simulated phylogeny
plot(sim_tree, main = "simulated phylogeny")
\donttest{
# find a solution that meets a budget of 300
s1 <- ppp_heuristic_solution(sim_project_data, sim_tree, 300,
                             "name", "cost", "success")

# print solution
print(s1)

# print the names of which projects were funded
print(names(s1)[which(unlist(s1[1, sim_project_data$name]))])

# plot solution
ppp_plot(sim_project_data, sim_tree, s1, "name", "cost", "success")

# find a solution that meets a budget of 300 and allocates
# funding for the "S1_project" project. For instance, species "S1" might
# be an iconic species that has cultural and economic importance.
sim_project_data2 <- sim_project_data
sim_project_data2$locked_in <- sim_project_data2$name == "S1_project"
s2 <- ppp_heuristic_solution(sim_project_data2, sim_tree, 300,
                             "name", "cost", "success",
                              locked_in_column_name = "locked_in")

# print solution
print(s2)

# plot solution
ppp_plot(sim_project_data2, sim_tree, s2, "name", "cost", "success")

# find a solution that meets a budget of 300 and does not allocate
# funding for the "S2_project" project. For instance, species "S2"
# might have very little cultural or economic importance. Broadly speaking,
# though, it is better to "lock in" "important" species rather than
# "lock out" unimportant species.
sim_project_data3 <- sim_project_data
sim_project_data3$locked_out <- sim_project_data2$name == "S2_project"
s3 <- ppp_heuristic_solution(sim_project_data3, sim_tree, 300,
                             "name", "cost", "success",
                             locked_out_column_name = "locked_out")

# print solution
print(s3)

# plot solution
ppp_plot(sim_project_data3, sim_tree, s3, "name", "cost", "success")

# find all solutions from the heuristic algorithm
# note we can set the budget higher than the total cost of all the
# projects, and the number of solutions to the total number of
# projects to achieve this
s4 <- ppp_heuristic_solution(sim_project_data, sim_tree,
                             sum(sim_project_data$cost) * 1.1,
                             "name", "cost", "success",
                              number_solutions = nrow(sim_project_data))

# print solutions
print(s4)

# plot solution cost against objective
plot(objective ~ cost, data = s4,
     main = "Heuristic solutions", xlab = "Cost ($)",
     ylab = "Expected phylogenetic diversity")
}
}
\references{
Bennett JR, Elliott G, Mellish B, Joseph LN, Tulloch AI,
Probert WJ, ... & Maloney R (2014) Balancing phylogenetic diversity
and species numbers in conservation prioritization, using a case study of
threatened species in New Zealand. \emph{Biological Conservation},
\strong{174}: 47--54.

Faith DP (2008) Threatened species and the potential loss of
phylogenetic diversity: conservation scenarios based on estimated extinction
probabilities and phylogenetic risk analysis. \emph{Conservation Biology},
\strong{22}: 1461--1470.

Joseph LN, Maloney RF, & Possingham HP (2009) Optimal
allocation of resources among threatened species: a project prioritization
protocol. \emph{Conservation biology}, \strong{23}, 328--338.

Rodrigues AS & Gaston KJ (2002) Optimisation in reserve selection
procedures---why not? \emph{Biological Conservation}, \strong{107}: 123-129.

Underhill LG (1994) Optimal and suboptimal reserve selection
algorithms. \emph{Biological Conservation}, \strong{70}: 85--87.
}
\seealso{
For other methods for generating solutions for the 'Project
  Prioritization Protocol' problem, see \code{\link{ppp_heuristic_solution}}
  \code{\link{ppp_exact_solution}}, and \code{\link{ppp_random_solution}}.
  To visualize the effectiveness of a particular solution, see
  \code{\link{ppp_plot}}.
}
