% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_manual_solution.R
\name{ppp_manual_solution}
\alias{ppp_manual_solution}
\title{Solve the 'Project Prioritization Protocol' problem by manually specifying
a solution}
\usage{
ppp_manual_solution(x, tree, budget, solution, project_column_name,
  cost_column_name, success_column_name, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tbl_df}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{verbose}{\code{logical} should information be printed while solving
the problem? No missing values are permitted. Defaults to \code{FALSE}.}
}
\value{
A \code{\code{link[tibble]{tbl_df}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{objective}}{\code{numeric} objective value associated with
     each of the solution(s). This corresponds to the expected amount of
     evolutionary history that is retained by each solution.}

   \item{\code{cost}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{optimal}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal or not.)}

 }}

 Finally, the object also contains following attributes
 (which can be accessed, for example, using \code{attr(output, "status")}):

 \describe{

   \item{\code{runtime}}{\code{numeric} number of seconds that elapsed
     while solving the problem.}

   \item{\code{status}}{\code{character} description of the
     solver status (e.g. \code{"OPTIMAL"} indicates that optimal solution(s)
     were found.)}

 }
}
\description{
Manually specify funding schemes for conservation projects and examine
their effectiveness.
}
\details{
TODO
}
\seealso{
The following methods can be used to extract, inspect, and
  visualize the solution(s): TODO
}
