% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_manual_solution.R
\name{ppp_manual_solution}
\alias{ppp_manual_solution}
\title{Solve the 'Project Prioritization Protocol' problem by manually specifying
a solution}
\usage{
ppp_manual_solution(x, tree, budget, solution, project_column_name,
  cost_column_name, success_column_name, verbose = FALSE)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{budget}{\code{numeric} value that represents the total budget available
for funding conservation projects.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{verbose}{\code{logical} should information be printed while solving
the problem? No missing values are permitted. Defaults to \code{FALSE}.}
}
\value{
A \code{\link[tibble]{tibble}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{objective}}{\code{numeric} objective value associated with
     each of the solution(s). This corresponds to the expected amount of
     evolutionary history that is retained by each solution.}

   \item{\code{cost}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{optimal}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal (\code{TRUE}) or not (\code{FALSE}.
     Missing values (\code{NA}) indicate that optimality is unknown
     (i.e. because the method used to produce the solution(s) does not
     provide any bounds on their quality).}

   \item{\code{method}}{\code{character} name of method used to produce the
     solution(s).)}

 }
}
\description{
Manually specify funding schemes for conservation projects and examine
their effectiveness. \strong{Since the solutions are manually specified,
their optimality is unknown.}.
}
\details{
TODO
}
\seealso{
For other methods for solving the 'Project Prioritization Protocol'
problem, see \code{\link{ppp_heuristic_solution}} and
\code{\link{ppp_gurobi_solution}}. To visualize the effectiveness of a
particular solution, see \code{\link{ppp_plot}}.
}
