% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_manual_solution.R
\name{ppp_manual_solution}
\alias{ppp_manual_solution}
\title{Solve the 'Project Prioritization Protocol' problem by manually specifying
a solution}
\usage{
ppp_manual_solution(x, tree, solution, project_column_name,
  cost_column_name, success_column_name)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{solution}{\code{\link[tibble]{tibble}} object containing the solution
data. Here, each row corresponds to a different solution. This object
should contain a column for each project---with the column name matching
the project name---and the column should contain \code{logical} values
that indicate if the project is funded (\code{TRUE}) or not (\code{FALSE}).
Although additional columns can also be included, they will be ignored.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{cost_column_name}{\code{character} name of column that
indicates the cost for funding each project. This column must have
\code{numeric} values which are equal to or greater than zero. No missing
values are permitted.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}
}
\value{
A \code{\link[tibble]{tibble}} object containing the
  solution(s) data. Each row corresponds to a different solution, and
  each column describes a different property of the solution. The object
  contains a column for each project (based on the argument to
  \code{project_column_name}) which contains \code{logical} values indicating
  if the project was prioritized for funded (\code{TRUE}) or not
  (\code{FALSE}) in a given solution. Additionally, the object also contains
  the following columns:

 \describe{

   \item{\code{objective}}{\code{numeric} objective value associated with
     each of the solution(s). This corresponds to the expected amount of
     evolutionary history that is retained by each solution.}

   \item{\code{cost}}{\code{numeric} total cost associated with each of
     of the solution(s).}

   \item{\code{optimal}}{\code{logical} indicating if each of the
     solution(s) is known to be optimal (\code{TRUE}) or not (\code{FALSE}.
     Missing values (\code{NA}) indicate that optimality is unknown
     (i.e. because the method used to produce the solution(s) does not
     provide any bounds on their quality).}

   \item{\code{method}}{\code{character} name of method used to produce the
     solution(s).)}

 }
}
\description{
Manually specify funding schemes for conservation projects and calculate
their effectiveness.
}
\examples{
# load built-in data
data(sim_project_data, sim_tree)

# load packages to help with plotting
library(ggplot2)

# print simulated project data set
print(sim_project_data)

# print simulated phylogenetic tree data set
print(sim_tree)

# plot the simulated phylogeny
plot(sim_tree, main = "simulated phylogeny")

# create some solutions, note that the column names the same as the values
# in the "name" column of the sim_project_data object
solutions <- data.frame(S1_project = c(FALSE, FALSE, TRUE),
                        S2_project = c(TRUE, FALSE, TRUE),
                        S4_project = c(TRUE, FALSE, TRUE),
                        S3_project = c(FALSE, FALSE, TRUE),
                        S5_project = c(TRUE, FALSE, TRUE),
                        baseline_project = c(TRUE, TRUE, TRUE))

print(solutions)

# evaluate the solutions
s1 <- ppp_manual_solution(sim_project_data, sim_tree, solutions,
                          "name", "cost", "success")

# print the output
print(s1)

# visualize the effectiveness of the different solutions
ppp_plot(sim_project_data, sim_tree, s1, "name", "cost", "success",
         n = 1) + ggtitle("solution 1")
ppp_plot(sim_project_data, sim_tree, s1, "name", "cost", "success",
         n = 2) + ggtitle("solution 2")
ppp_plot(sim_project_data, sim_tree, s1, "name", "cost", "success",
         n = 3) + ggtitle("solution 3")
}
\seealso{
For other methods for generating solutions for the 'Project
  Prioritization Protocol' problem, see \code{\link{ppp_gurobi_solution}},
  \code{\link{ppp_heuristic_solution}}, and
  \code{\link{ppp_random_solution}}. To visualize the effectiveness of a
  particular solution, see \code{\link{ppp_plot}}.
}
