% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ppp_objective_value}
\alias{ppp_objective_value}
\title{'Project Prioritization Protocol' objective value}
\usage{
ppp_objective_value(x, tree, project_column_name, success_column_name,
  solution)
}
\arguments{
\item{x}{\code{\link[base]{data.frame}} or \code{\link[tibble]{tibble}}
table containing project data. Here, each row should correspond to
a different project and columns should contain that correspond to
each project. See the Details section below for more information.}

\item{tree}{\code{\link[ape]{phylo}} phylogenetic tree describing the
evolutionary history of the species affected by the conservation
projects that could potentially be funded. Note that every single
species that is affected by the various conservation projects should
be represented in this tree.}

\item{project_column_name}{\code{character} name of column that contains
the name for each conservation project. Note that the project names
must not contain any duplicates or missing values.}

\item{success_column_name}{\code{character} name of column that
denotes the probability that each project will succeed if it is funded.
This column must have \code{numeric} values which lay between zero and one.
No missing values are permitted.}

\item{solution}{\code{\link[tibble]{tibble}} indicating if each project
is funded or not. Here, each column corresponds to a different project and
each row corresponds to a different solution.}
}
\value{
\code{numeric} vector with an objective value for each solution
 in the argument to \code{solutions}.
}
\description{
Calculate the objective value for a set of solutions to a Project
Prioritization Protocol problem.
}
\keyword{internal}
