% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppp_simulate_data.R
\name{ppp_simulate_data}
\alias{ppp_simulate_data}
\title{Simulate conservation project data}
\usage{
ppp_simulate_data(number_species, cost_mean = 100, cost_sd = 5,
  success_min_probability = 0.7, success_max_probability = 0.99,
  funded_min_persistence_probability = 0.5,
  funded_max_persistence_probability = 0.9,
  not_funded_min_persistence_probability = 0.01,
  not_funded_max_persistence_probability = 0.4,
  locked_in_proportion = 0, locked_out_proportion = 0)
}
\arguments{
\item{number_species}{\code{numeric} number of species.}

\item{cost_mean}{\code{numeric} average cost for the projects. Defaults to
\code{100}.}

\item{cost_sd}{\code{numeric} standard deviation in project costs. Defaults
to \code{5}.}

\item{success_min_probability}{\code{numeric} minimum probability of the
projects succeeding if they are funded. Defaults to \code{0.7}.}

\item{success_max_probability}{\code{numeric} maximum probability of the
projects succeeding if they are funded. Defaults to \code{0.99}.}

\item{funded_min_persistence_probability}{\code{numeric} minimum probability
of the species persisting if their projects are funded and successful.
Defaults to \code{0.5}.}

\item{funded_max_persistence_probability}{\code{numeric} maximum probability
of the species persisting if their projects are funded and successful.
Defaults to \code{0.9}.}

\item{not_funded_min_persistence_probability}{\code{numeric} minimum
probability of the species persisting if their projects are not funded.
Defaults to \code{0.01}.}

\item{not_funded_max_persistence_probability}{\code{numeric} maximum
probability of the species persisting if their projects are not funded.
Defaults to \code{0.4}.}

\item{locked_in_proportion}{\code{numeric} of projects that are locked
into the solution. Defaults to \code{0}.}

\item{locked_out_proportion}{\code{numeric} of projects that are locked
into the solution. Defaults to \code{0}.}
}
\value{
A \code{list} object containing the elements:

  \describe{

    \item{\code{"project_data"}}{A \code{\link[tibble]{tibble}} containing
      the data for the conservation projects. It contains the following
      columns:

      \describe{


        \item{\code{"success"}}{\code{numeric} probability of each project
          succeeding if it is funded.}

        \item{\code{"S1"} ... \code{"SN"}}{\code{numeric} columns for each
          species, ranging from \code{"S1"} to \code{"SN"} where \code{N}
          is the number of species, indicating the enhanced probability that
          each species will persist if it funded.}

        \item{\code{"S1_action"} ... \code{"SN_action"}}{\code{logical}
          columns for each action, ranging from \code{"S1_action"} to
          \code{"SN_action"} where \code{N} is
          the number of actions (equal to the number of species in this
          simulated data), indicating if an action is associated with a
          project (\code{TRUE}) or not (\code{FALSE}).}

    }}

    \item{\code{"action_data"}}{A \code{\link[tibble]{tibble}} containing
      the data for the conservation actions. It contains the following
      columns:

      \describe{


        \item{\code{"cost"}}{\code{numeric} cost for each action.}

        \item{\code{"locked_in"}}{\code{logical} indicating if certain
          actions should be locked into the solution.}

        \item{\code{"locked_out"}}{\code{logical} indicating if certain
          actions should be locked out of the solution.}

    }}

   \item{tree}{\code{\link[ape]{phylo}} phylogenetic tree for the species.}

 }
}
\description{
Simulate conservation project data to develop simulated prioritizations.
}
\details{
The simulated data set will contain one conservation project for
  each species and a "baseline" (do nothing) project to reflect species'
  persistence when none of their conservation projects are not funded. Each
  conservation project is associated with a single action, and no
  conservation projects share any actions. Specifically, the data are
  simulated as follows:

  \enumerate{

    \item A conservation project is created for each species, and each
      project is associated with its own single action.

    \item Cost data for each action are simulated using a normal
      distribution and the \code{cost_mean} and \code{cost_sd} arguments.

    \item A set proportion of the actions are randomly set to be locked
      in and out of the solutions using the \code{locked_in_proportion} and
      \code{locked_out_proportion} arguments.

    \item The probability of each project succeeding if its action is funded
      is simulated by drawing probabilities from a uniform distribution with
      the upper and lower bounds set as the \code{success_min_probability}
      and \code{success_max_probability} arguments.

    \item The probability of each species persisting if its project is funded
      and is successful is simulated by drawing probabilities from a uniform
      distribution with the upper and lower bounds set as the
      \code{funded_min_persistence_probability} and
      \code{funded_max_persistence_probability} arguments.

    \item An additional project is created which represents the "baseline"
      (do nothing) scenario. The probability of each species persisting
      when managed under this project is simulated by drawing probabilities
      from a uniform distribution with the upper and lower bounds
      set as the \code{not_funded_min_persistence_probability}
      and \code{not_funded_max_persistence_probability} arguments.

    \item A phylogenetic tree is simulated for the species using
      \code{\link[ape]{rcoal}}.

 }
}
\examples{
# create a simulated data set
s <- ppp_simulate_data(number_species = 5,
                         cost_mean = 100,
                         cost_sd = 5,
                         success_min_probability = 0.7,
                         success_max_probability = 0.99,
                         funded_min_persistence_probability = 0.5,
                         funded_max_persistence_probability = 0.9,
                         not_funded_min_persistence_probability = 0.01,
                         not_funded_max_persistence_probability = 0.4,
                         locked_in_proportion = 0.01,
                         locked_out_proportion = 0.01)

# print data set
print(s)

}
